Update sr-asset executor

+++ libs/sr/assets/src/executors/gen-assets/executor.ts
@@ -1,21 +1,7 @@
 import { generateIndexFromObj } from '@genshin-optimizer/common/pipeline'
-import { crawlObject } from '@genshin-optimizer/common/util'
-import type {
-  CharacterKey,
-  LightConeKey,
-  RelicCavernSetKey,
-  RelicPlanarSetKey,
-} from '@genshin-optimizer/sr/consts'
-import {
-  DM2D_PATH,
-  avatarConfig,
-  characterIdMap,
-  equipmentConfig,
-  lightConeIdMap,
-  relicDataInfo,
-  relicSetIdMap,
-  relicSlotMap,
-} from '@genshin-optimizer/sr/dm'
+import { crawlObject, layeredAssignment } from '@genshin-optimizer/common/util'
+import { AssetData } from '@genshin-optimizer/sr/assets-data'
+import { DM2D_PATH } from '@genshin-optimizer/sr/dm'
 import { workspaceRoot } from '@nx/devkit'
 import * as fs from 'fs'
 import * as path from 'path'
@@ -23,42 +9,9 @@ import type { GenAssetsExecutorSchema } from './schema'
 
 const DEST_PROJ_PATH = `${workspaceRoot}/libs/sr/assets/src` as const
 
-type CharacterIcon = {
-  icon: string
-  cover: string
-}
-
-type LightConeIcon = {
-  icon: string
-  cover: string
-}
-
-type RelicPlanarIcons = Record<
-  RelicPlanarSetKey,
-  { rope: string; sphere: string }
->
-type RelicCavernIcons = Record<
-  RelicCavernSetKey,
-  { head: string; sphere: string }
->
-type RelicIcons = RelicPlanarIcons & RelicCavernIcons
-//An object to store all the asset related data.
-export const AssetData = {
-  // artifacts: {},
-  lightCones: {} as Record<LightConeKey, LightConeIcon>,
-  chars: {} as Record<CharacterKey, CharacterIcon>,
-  relic: {} as RelicIcons,
-}
-
 export default async function runExecutor(
   _options: GenAssetsExecutorSchema
 ): Promise<{ success: boolean }> {
-  // Best effort and silently fail since most of the time we don't use this
-
-  if (!fs.existsSync(DM2D_PATH)) {
-    console.log(`\`assets\` folder does not exist, no assets will be copied.`)
-    return { success: true }
-  }
   function copyFile(src: string, dest: string) {
     if (!fs.existsSync(src)) {
       console.warn('Cannot find file', src)
@@ -70,59 +23,38 @@ export default async function runExecutor(
     })
   }
 
-  // Get icons for each artifact piece
-  AssetData.relic = Object.fromEntries(
-    Object.entries(relicDataInfo).map(([relicSetId, reflicDatas]) => [
-      relicSetIdMap[relicSetId],
-      Object.fromEntries(
-        Object.entries(reflicDatas).map(([relicSlotKey, relicData]) => [
-          relicSlotMap[relicSlotKey],
-          relicData.ItemFigureIconPath,
-        ])
-      ),
-    ])
-  ) as RelicIcons
-
-  // parse baseStat/ascension/basic data for characters.
-  Object.entries(avatarConfig).forEach(([charid, charData]) => {
-    const { DefaultAvatarHeadIconPath, AvatarCutinFrontImgPath } = charData
-
-    const assets: CharacterIcon = {
-      icon: DefaultAvatarHeadIconPath.toLocaleLowerCase(),
-      cover: AvatarCutinFrontImgPath.toLocaleLowerCase(),
-    }
-    AssetData.chars[characterIdMap[charid]] = assets
-  })
-
-  Object.entries(equipmentConfig).forEach(([id, data]) => {
-    const { ThumbnailPath, ImagePath } = data
-
-    const assets: LightConeIcon = {
-      icon: ThumbnailPath.toLocaleLowerCase(),
-      cover: ImagePath.toLocaleLowerCase(),
-    }
-    AssetData.lightCones[lightConeIdMap[id]] = assets
-  })
-
   // Dump out the asset List.
-
-  // dumpFile(
-  //   `${WORKSPACE_ROOT_PATH}/libs/sr/assets/src/lib/AssetData_gen.json`,
-  //   AssetData
-  // )
+  // Best effort and since most of the time we don't use this
+  if (fs.existsSync(DM2D_PATH)) {
+    console.log('Copying files')
+    crawlObject(
+      AssetData,
+      [],
+      (s) => typeof s === 'string',
+      (filePath: string, keys) => {
+        const fileName = keys.slice(-1)
+        const folderPath = keys.slice(0, -1).join('/')
+        copyFile(
+          `${DM2D_PATH}/${filePath}`,
+          `${DEST_PROJ_PATH}/gen/${folderPath}/${fileName}.png`
+        )
+      }
+    )
+  } else {
+    console.log(`\`assets\` folder does not exist, no assets will be copied.`)
+  }
+  console.log('Generating index')
+  // Extract just the filename for the index files
+  const indexData = {}
   crawlObject(
     AssetData,
     [],
     (s) => typeof s === 'string',
-    (icon, keys) => {
-      copyFile(
-        `${DM2D_PATH}/${icon}`,
-        `${DEST_PROJ_PATH}/gen/${keys.slice(0, -1).join('/')}/${keys.at(
-          -1
-        )}.png`
-      )
+    (_filePath: string, keys) => {
+      const fileName = keys.slice(-1)[0]
+      layeredAssignment(indexData, keys, fileName)
     }
   )
-  generateIndexFromObj(AssetData, `${DEST_PROJ_PATH}/gen`)
+  await generateIndexFromObj(indexData, `${DEST_PROJ_PATH}/gen`)
   return { success: true }
 }
